Fix list file
