+++ query/src/interpreters/interpreter_list.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::io;
 use std::sync::Arc;
 
 use common_datablocks::DataBlock;
@@ -19,11 +20,11 @@ use common_datavalues::Series;
 use common_datavalues::SeriesFrom;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_io::prelude::operator_list_files;
 use common_planners::ListPlan;
 use common_streams::DataBlockStream;
 use common_streams::SendableDataBlockStream;
 use common_tracing::tracing;
+use futures::StreamExt;
 use regex::Regex;
 
 use crate::interpreters::Interpreter;
@@ -45,7 +46,23 @@ impl ListInterpreter {
         let op = StageSource::get_op(&self.ctx, &self.plan.stage).await?;
         let pattern = &self.plan.pattern;
         let path = &self.plan.path;
-        let mut files = operator_list_files(&op, path).await?;
+
+        let mut files = if path.ends_with('/') {
+            let mut list = vec![];
+            let mut objects = op.object(path).list().await?;
+            while let Some(object) = objects.next().await {
+                let name = object?.name();
+                list.push(name);
+            }
+            list
+        } else {
+            let o = op.object(path);
+            match o.metadata().await {
+                Ok(_) => vec![o.name()],
+                Err(e) if e.kind() == io::ErrorKind::NotFound => vec![],
+                Err(e) => return Err(e.into()),
+            }
+        };
 
         if !pattern.is_empty() {
             let regex = Regex::new(pattern).map_err(|e| {
